+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v2.3.0
+------------------------------
+*October 5, 2020*
+
+### Added
+- Axe packages for running a11y tests in any component.
+
+
+v2.2.0
+------------------------------
+*September 29, 2020*
+
+### Added
+- Bundlewatch config for monitoring bundle file sizes.
+
+
 v2.1.0
 ------------------------------
 *September 28, 2020*
@@ -20,12 +36,12 @@ v2.0.0
 
 
 v1.34.0
-*September 24, 2020*
 ------------------------------
+*September 24, 2020*
+
 ### Changed
 - Lerna script to run component tests in series.
 - `maxInstanstances` property in WebDriverIO config
-
 
 ### Removed
 - Contract test dependency.
@@ -34,8 +50,9 @@ v1.34.0
 
 
 v1.33.0
-*September 21, 2020*
 ------------------------------
+*September 21, 2020*
+
 ### Removed
 - Contract test dependencies + scripts
 
@@ -283,6 +300,7 @@ v1.10.0
 
 ### Added
 - NPM Scripts at the root of the mono repo to run the Storybook docs through `yarn storybook:build` and `yarn storybook:serve`.
+
 
 v1.9.0
 ------------------------------